moved send.go into separate utils for loaders
