+++ internal/controller/discovery/loaders/utils/send.go
@@ -1,4 +1,4 @@
-package loaders
+package utils
 
 import (
 	"context"
